Extract DeliusServiceUser serialization function
into own class.

This was doing quite a lot in the `interventionsService` now that we've
added the `ContactDetails` to it. This separation feels cleaner to me.

This commit also fixes an issue with some nullable fields that arent
present in the response from the CommunityAPI (at least locally). This
issue came to light when removing unneccessary type assertions on the
initial implementations, and became hard to separate into its own
commit.
